ML/requirements.txt
absl-py
asttokens
astunparse
attrs
beautifulsoup4
blinker
bs4
certifi
cffi
charset-normalizer
click
colorama
comm
contourpy
cycler
debugpy
decorator
executing
filelock
Flask
flatbuffers
fonttools
fsspec
gast
google-pasta
grpcio
h11
h5py
icrawler
idna
ipykernel
ipython
ipython_pygments_lexers
itsdangerous
jedi
Jinja2
joblib
jupyter_client
jupyter_core
keras
kiwisolver
libclang
lxml
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
ml_dtypes
mpmath
namex
nest-asyncio
networkx
numpy
opencv-python
opt_einsum
optree
outcome
packaging
pandas
parso
pillow
platformdirs
prompt_toolkit
protobuf
psutil
pure_eval
py-cpuinfo
pycparser
Pygments
pyparsing
PySocks
python-dateutil
python-dotenv
pytz
PyYAML
pyzmq
requests
rich
scikit-learn
scipy
seaborn
selenium
six
sniffio
sortedcontainers
soupsieve
stack-data
sympy
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl
torch
torchvision
tornado
tqdm
traitlets
trio
trio-websocket
typing_extensions
tzdata
ultralytics
ultralytics-thop
urllib3
wcwidth
webdriver-manager
websocket-client
Werkzeug
wrapt
wsproto
